+++ commodore/postprocess/builtin_filters.py
@@ -10,7 +10,7 @@ from commodore import __install_dir__
 from .jsonnet import jsonnet_runner
 
 
-def _builtin_filter_helm_namespace(inv, component, target, path, **kwargs):
+def _builtin_filter_helm_namespace(inv, component, path, **kwargs):
     if "namespace" not in kwargs:
         raise click.ClickException(
             "Builtin filter 'helm_namespace': filter argument 'namespace' is required"
@@ -18,13 +18,12 @@ def _builtin_filter_helm_namespace(inv, component, target, path, **kwargs):
     create_namespace = kwargs.get("create_namespace", "false")
     exclude_objects = kwargs.get("exclude_objects", [])
     exclude_objects = "|".join([json.dumps(e) for e in exclude_objects])
-    output_dir = P("compiled", target, path)
+    output_dir = P("compiled", component, path)
 
     # pylint: disable=c-extension-no-member
     jsonnet_runner(
         inv,
         component,
-        target,
         path,
         _jsonnet.evaluate_file,
         __install_dir__ / "filters" / "helm_namespace.jsonnet",
@@ -40,9 +39,9 @@ _builtin_filters = {
 }
 
 
-def run_builtin_filter(inv, component, target, f):
+def run_builtin_filter(inv, component, f):
     fname = f["filter"]
     if fname not in _builtin_filters:
         click.secho(f"   > [ERR ] Unknown builtin filter {fname}", fg="red")
         return
-    _builtin_filters[fname](inv, component, target, f["path"], **f["filterargs"])
+    _builtin_filters[fname](inv, component, f["path"], **f["filterargs"])