+++ tests/test_target.py
@@ -37,14 +37,21 @@ def cluster_from_data(data) -> cluster.Cluster:
 
 def test_render_target(tmp_path):
     os.chdir(tmp_path)
-    bar_defaults = P("inventory/classes/defaults/bar.yml")
-    os.makedirs(bar_defaults.parent, exist_ok=True)
-    bar_defaults.touch()
-    target = cluster.render_target("foo", ["bar", "baz"])
+    classdir = P("inventory/classes")
+    for cls in ["foo", "bar"]:
+        defaults = classdir / "defaults" / f"{cls}.yml"
+        os.makedirs(defaults.parent, exist_ok=True)
+        defaults.touch()
+        component = classdir / "components" / f"{cls}.yml"
+        os.makedirs(component.parent, exist_ok=True)
+        component.touch()
+    target = cluster.render_target("foo", ["foo", "bar", "baz"])
     classes = [
-        "params.foo",
+        "params.cluster",
+        "defaults.foo",
         "defaults.bar",
         "global.commodore",
+        "components.foo",
     ]
     assert target != ""
     print(target)
@@ -53,18 +60,19 @@ def test_render_target(tmp_path):
     ), "rendered target includes different amount of classes"
     for i in range(len(classes)):
         assert target["classes"][i] == classes[i]
+    assert target["parameters"]["kapitan"]["vars"]["target"] == "foo"
 
 
 def test_render_params(data):
-    params = cluster.render_params(cluster_from_data(data), "foo")
-    assert params["parameters"]["target_name"] == "foo"
+    params = cluster.render_params(cluster_from_data(data))
     assert params["parameters"]["cluster"]["name"] == "mycluster"
+    assert params["parameters"][cluster.BOOTSTRAP_TARGET]["name"] == "mycluster"
     assert (
-        params["parameters"]["cluster"]["catalog_url"]
+        params["parameters"][cluster.BOOTSTRAP_TARGET]["catalog_url"]
         == "ssh://git@git.example.com/cluster-catalogs/mycluster"
     )
-    assert params["parameters"]["cluster"]["tenant"] == "mytenant"
-    assert params["parameters"]["cluster"]["dist"] == "rancher"
+    assert params["parameters"][cluster.BOOTSTRAP_TARGET]["tenant"] == "mytenant"
+    assert params["parameters"][cluster.BOOTSTRAP_TARGET]["dist"] == "rancher"
     assert params["parameters"]["facts"] == data["facts"]
     assert params["parameters"]["cloud"]["provider"] == "cloudscale"
     assert params["parameters"]["customer"]["name"] == "mytenant"
@@ -73,18 +81,18 @@ def test_render_params(data):
 def test_missing_facts(data):
     data["facts"].pop("cloud")
     with pytest.raises(click.ClickException):
-        cluster.render_params(cluster_from_data(data), "foo")
+        cluster.render_params(cluster_from_data(data))
 
 
 def test_empty_facts(data):
     data["facts"]["cloud"] = ""
     with pytest.raises(click.ClickException):
-        cluster.render_params(cluster_from_data(data), "foo")
+        cluster.render_params(cluster_from_data(data))
 
 
 def test_read_cluster_and_tenant(tmp_path):
     os.chdir(tmp_path)
-    file = cluster.params_file("foo")
+    file = cluster.params_file()
     os.makedirs(file.parent, exist_ok=True)
     with open(file, "w") as f:
         f.write(
@@ -97,14 +105,14 @@ def test_read_cluster_and_tenant(tmp_path):
             )
         )
 
-    cluster_id, tenant_id = cluster.read_cluster_and_tenant("foo")
+    cluster_id, tenant_id = cluster.read_cluster_and_tenant()
     assert cluster_id == "c-twilight-water-9032"
     assert tenant_id == "t-delicate-pine-3938"
 
 
 def test_read_cluster_and_tenant_missing_fact(tmp_path):
     os.chdir(tmp_path)
-    file = cluster.params_file("foo")
+    file = cluster.params_file()
     os.makedirs(file.parent, exist_ok=True)
     with open(file, "w") as f:
         f.write(
@@ -116,4 +124,4 @@ def test_read_cluster_and_tenant_missing_fact(tmp_path):
         )
 
     with pytest.raises(KeyError):
-        cluster.read_cluster_and_tenant("foo")
+        cluster.read_cluster_and_tenant()